pyproject.toml
[tool.poetry]
name = "fastapi-sllm-api"
version = "0.1.0"
description = "FastAPI 기반 SLLM 바이너리 API 서버"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
pydantic = "^2.4.2"
python-dotenv = "^1.0.0"
httpx = "^0.25.0"
python-multipart = "^0.0.6"
llama-cpp-python = "^0.3.8"
pydantic-settings = "^2.8.1"

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
black = "^23.10.1"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
